examples/requirements.txt
numpy==1.19.5
spidev==3.5
Pillow==5.4.1
RPi.GPIO==0.7.0
